move util into class method
