+++ packages/facility-server/app/routes/apiv1/index.js
@@ -48,7 +48,6 @@ import { vitals } from './vitals';
 import { template } from './template';
 import { translation } from './translation';
 import { vaccinationSettings } from './vaccinationSettings';
-import { getLanguageOptions } from '@tamanu/shared/utils/translation/getLanguageOptions';
 
 export const apiv1 = express.Router();
 const patientDataRoutes = express.Router();
@@ -70,8 +69,9 @@ apiv1.get(
 
 apiv1.get('/public/translation/preLogin', async (req, res) => {
   req.flagPermissionChecked();
-  const response = await getLanguageOptions(req.models);
-  res.send(response.languageOptions);
+  const { TranslatedString } = req.models;
+  const response = await TranslatedString.getPossibleLanguages();
+  res.send(response);
 });
 
 apiv1.get(
