+++ packages/web/app/api/queries/useTranslationLanguages.js
@@ -1,7 +1,7 @@
 import { useQuery } from '@tanstack/react-query';
 import { useApi } from '../useApi';
 
-export const useTranslationOptions = () => {
+export const useTranslationLanguages = () => {
   const api = useApi();
   return useQuery(['languageList'], () => api.get('public/translation/preLogin'));
 };
